Downloads/Magnus-Deal-Scout-Agent/files (5)/backend-requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Celery for Background Tasks
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# HTTP Requests & Web Scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2
playwright==1.41.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
shortuuid==1.0.11

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# CORS
python-cors==1.0.0
